source/microservices/product_service/requirements.txt
Flask 
flask-cors 
mysql-connector-python 
bcrypt